server/requirements.txt
fastapi
uvicorn
pydantic
pyyaml
httpx
openai
gradio
openenv-core>=0.2.0
